Added fetch to get random photo and reestructuring the directories
