+++ src/pages/SearchPage.js
@@ -2,12 +2,11 @@ import { useEffect, useState } from "react";
 import { Container } from "@mui/material";
 import { InputSearch } from "../components/TextField";
 import { Spinner } from "../components/spinner";
-import { conversionDataApi } from "../components/conversion-data-from-api";
 import { GalleryImages } from "../components/image-list";
 import { useDebounce } from "../custom-hooks/useDebounce";
 import { addPhoto } from "../features/my-photos/myPhotosSlice";
 import { useDispatch } from "react-redux";
-
+import { searchCharacters } from "../services/unsplash-api";
 export const Search = () => {
 	// https://usehooks.com/useDebounce/
 	const [searchTerm, setSearchTerm] = useState("");
@@ -57,24 +56,3 @@ export const Search = () => {
 		</>
 	);
 };
-
-function searchCharacters(search) {
-	const API_KEY = "MIMdH3XPFMcOFvYg9cbiQ5iLuiLml2Fa14CGidU5ZXM";
-	// TODO Serían valores que se podrían modificar y se podrían pasar por parámetros
-	const CANTIDAD_POR_PAGINA = 25;
-	const PAGE = 1;
-	return fetch(`https://api.unsplash.com/search/photos/?query=${search}&page=${PAGE}&per_page=${CANTIDAD_POR_PAGINA}&client_id=${API_KEY}`,
-		{
-			method: "GET",
-		}
-	)
-		.then((response) => response.json())
-		.then((response) => {
-			const resultApi = response.results.map(item => conversionDataApi(item));
-			return resultApi;
-		})
-		.catch((error) => {
-			console.error(error);
-			return [];
-		});
-}
